+++ NEWS.md
@@ -1,3 +1,7 @@
+# tarflow.iquizoo 3.12.7
+
+* Internal upkeeps ([#100](https://github.com/iquizoo/tarflow.iquizoo/issues/100))
+
 # tarflow.iquizoo 3.12.6
 
 ## Bug Fixes